Implementation of the algorithm -
With the basic functions developed in the previous commits, it was
possible to start the implementation of the ARS. I created 3 new
attributes for the graph: a list of existing colors, a list of deleted
colors and a graph (it is used as a back up, in case the algorithm goes
wrong). The changeColor, adaptNeighbours, prepareBackUp, chargeBackUp,
getRandomExistingColors and eraseVertexColor functions have been
created.